refactor: moved product_service
