+++ product_service/product_service/dynamo_db_stack.py
