+++ product_service/product_service/lambda_functions/create_product.py
