+++ product_service/product_service/lambda_functions/get_product_by_id.py
