+++ product_service/product_service/lambda_functions/get_products_list.py
