+++ product_service/product_service/lambda_functions/utils/dynamodb.py
