+++ product_service/product_service/lambda_functions/utils/logging.py
@@ -1,7 +1,7 @@
 import json
 import logging
 
-logger_request = logging.getLogger("aws_shop_serverless")
+logger_request = logging.getLogger("product_service")
 logger_request.setLevel(logging.INFO)
 logger_request.addHandler(logging.StreamHandler())
 
@@ -10,7 +10,7 @@ def log_request(event):
     """Log request (Lambda event) data:
     - body
     - path params"""
-    
+
     try:
         body = json.dumps(event.get("body", {}), indent=2)
         path_params = json.dumps(event.get("pathParameters", {}), indent=2)
