+++ product_service/product_service/utils/populate_dynamodb.py
